feat: add class RenderTarget
Refined class Screen, DesktopScreen and WindowsConsoleScreen with render targets.
Now they can act as a sink of the rendering pipeline 🥳!

Original interfaces like WritePixel and ReadPixel are kept, which means
that we can still directly manipulate the pixels, thus no need to modify
the demo codes.

+++ demo/LifeGame/main.cpp
@@ -41,7 +41,7 @@ int main(int argc, char *argv[]) {
 					for (int j = -1; j <= 1; j++) {
 						int nrow = static_cast<int>(row) + i;
 						int ncol = static_cast<int>(col) + j;
-						if (nrow >= 0 && nrow < height && ncol >= 0 && ncol < width) {
+						if (nrow >= 0 && nrow < static_cast<int>(height) && ncol >= 0 && ncol < static_cast<int>(width)) {
 							alive_cnt += cells_prev[nrow][ncol];
 						}
 					}
